Moved DHT plugin to ethpool framework
